make: depends.d -> .deps/%.d
Each invocation of the compiler outputs dependency information for just
that file as a side effect.

This eliminates the initial call to `gcc -MM` at the beginning of every
invocation of `make`.

If `%.c`'s `.deps/%.d` is missing, its `%.o` probably doesn't exist
either, because if `%.o` had been generated, so would have `.deps/%.d`.
Also, if a `.deps/%.d` is outdated because `%.c` or something included
by it had includes added, it still won't be a problem, because `%.o` has
to be rebuilt anyway, because the file changed regardless of the fact
that more includes were added.  Therefore, nothing that should be
rebuilt will ever not be rebuilt.