Replace nest argument from aem_stream_provide_begin with currently-unused force argument
Provide nesting is now always permitted---check for it yourself if you
need to guard against it.

The force argument is currently ignored, but the intention is to make
aem_stream_provide_begin indicate whether aem_stream_should_provide
would return false if the provide isn't nested and the force flag isn't
set.