Warn if aem_logf format string ends with a newline
The old way of including a newline in aem_logf format strings is hereby
deprecated.  Currently, the right thing will happen whether or not a
message ends in a newline, but the check will eventually be removed,
causing messages that supply their own newlines to be followed by an
empty line.