utf8: rewrite as table-driven, remove ambiguity between 0xFFFFFFFF and EOF
aem_stringbuf_put and aem_stringslice_{get => get_rune} were rewritten
to be table-driven.  gcc-11.2.0 -O3 compiles away the tables, but
clang-13 -O3 doesn't seem to.

aem_stringslice_get uses a return value of -1 to indicate both
0xFFFFFFFF and failure.  This commit adds a new aem_stringslice_get_rune
which returns zero iff it fails to get a codepoint, and which accepts
a pointer to a uint32_t indicating where it should store the result.

The new aem_stringslice_match_rune function is a wrapper for
aem_stringslice_get_rune that returns a stringslice indicating the
matched bytes (or, on failure, an empty slice at the input position).

The old aem_stringslice_get function is now a wrapper for
aem_stringslice_get_rune, and may become deprecated in the future.